lb: fix handling flag `loadbalancer-cp-accesslog-enable-tcp`
Currently, the flag `loadbalancer-cp-accesslog-enable-tcp` isn't mapped
properly  which results in the TCP access log on T2 always being enabled.

This commit fixes the mapping - after which the default is back to TCP
access logging not being disabled (as it's quite verbose in combination
with the T1->T2 health checks.

Note: It's still enabled in development via `kind-loadbalancer.yaml`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>