lb: use proper access log filter to filter out T1->T2 HC requests
Currently, the option to filter out T1->T2 health check requests from
the T2 HTTP access log is implemented by using a header based access
log filter that we configure to exclude requests with the according
user agent.

Since Envoy already provides a built-in filter for this purpose, this
commit replaces the header based filter with the built-in one.

Note: In addition, the config value has been added to the file
`kind-loadbalancer.yaml`. Even though still filtering the HC requests
(which is also the default) - it allows for easy change of this config.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>