egressgw: standalone mode agent implementation
[ v1.15-ce commit 9d10f6709573d40363280cf3e86e9820519f57d2 ]

[ backporter's notes: aligned with new hive/cell API
  applied 2faab0e29408d36f2284c7fe73dc82dd2d46320c to port SEGW
  to v2 policy map ]

Standalone Egress Gateway (SEGW) represents a variant of standard
Egress Gateway designed to be deployed on nodes outside of a K8s
clusters. It is responsible for handling the traffic coming in
from the VXLAN tunnel, redirecting it to the configured egress
interface and masquerading it accordingly.

This commit implements the agent side changes to configure the
standalone egress gateway functionality. Specifically, it enables
tunneling support and configures the appropriate datapath macros
to trigger the compilation of the relevant code paths. Additionally,
it configures a default egress policy that matches all traffic
coming in from the tunnel, and masquerades it to the first address
of the interface specified as parameter (or the one with the
default route, if none is specified).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>