ci: don't restart ds/cilium after ds/cilium-dnsproxy upgrade
There is no need to restart cilium at this point. Only the
cilium-dnsproxy was upgraded and it should be able to talk to
cilium-agent which is already running. Because cilium was just restarted
one step before and is just coming up again, it might even cause issues
when cilium-dnsproxy is restoring rules but cilium-agent is going down
again in the meantime.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>