egressgw: standalone mode datapath implementation
[ v1.15-ce commit f959d8cce22c9c58aa00e411d0022f298afa38a6 ]

[ backporter's notes: add the SEGW logic to
  egress_gw_reply_needs_redirect_hook because the
  commit 436f7404adaa65f81935cb2809d4928f922bed7c split up
  the EgressGW policy lookup, and the subsequent ipcache lookup ]

Standalone Egress Gateway (SEGW) represents a variant of standard
Egress Gateway designed to be deployed on nodes outside of a K8s
clusters. It is responsible for handling the traffic coming in
from the VXLAN tunnel, redirecting it to the configured egress
interface and masquerading it accordingly.

This commit implements the core datapath changes to learn the
endpoint IP to tunnel endpoint mappings, as well as to handle
reply traffic. Specifically, the mappings are learned in the
from_overlay path, retrieving the remote tunnel endpoint (and
security identity) from the tunnel metadata, and stored into
the dedicated map. The reply path is modified accordingly to
lookup the tunnel endpoint from the previously populated
cilium_egress_gw_standalone_v4 map, rather than from the
ipcache, as performed normally.

The new logic depends on the standard egress gateway HA machinery,
and is additionally guarded by the ENABLE_EGRESS_GATEWAY_STANDALONE
macro.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>