enterprise/encryption: Add encryption policy computation metrics
This adds two new encryption policy metrics, measuring the number of
times the policy computation was triggered and how long the invocation
took on average.

Notably, this does not include the time it takes to push down the
computed tuples down the BPF map. Since that happens out of band in the
reconciler, separate metrics are used to make that observable.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>