lb: cleanup toTLSCertificateSdsConfigs
Currently, the translation from the TLSConfig into Envoy SDSConfigs
assumes that there can only be one application enabled per LBService.

This is true for the moment, but might change in the future.

Therefore, this commit changes the logic of the function to retrieve
the TLSConfig from the caller. This way, it's not up to the logic of
the function to decide from which application the TLS secrets need
to be mapped from.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>