gha: standalone egress gateway test
[ v1.15-ce commit 88462981541db3ecfb4a423485c052114318ccdb ]

[ backporter's notes: hit minor conflicts due to different surrounding
  context, adapted as appropriate. ]

Introduce a Github Action workflow to validate the standalone egress
gateway functionality.

At high level, the workflow initially sets up a Kind cluster, which is
connected to the internal network, and deploys Cilium, configured in
native routing mode (to simulate ENI mode), and with egress gateway HA
enabled. Additionally, it spawns two gateway containers, attached to
both the internal and the external network, running Cilium in standalone
egress gateway mode. Synthetic CiliumNode objects are manually created
on the workload cluster to represent the gateway nodes. Finally, it
deploys the external app connected to the external network only.

The overall setup is depicted in the following:

    ┌──────────────────┐
    │                  │ INT               EXT
    │ WORKLOAD CLUSTER │  │  ┌───────────┐  │
    │                  │  │  │           │  │
    │ ┌──────────────┐ │  │  │           │  │   ┌─────┐
    │ │              │ │  │  │  GATEWAY  ┼──┼   │     │
    │ │              ┼─┼──┼  │           │  │   │ APP │
    │ └──────────────┘ │  ┼──┼           │  ┼───┼     │
    │ ┌──────────────┐ │  │  └───────────┘  │   └─────┘
    │ │              │ │  │  ┌───────────┐  │
    │ │              ┼─┼──┼  │           │  │   ┌─────┐
    │ └──────────────┘ │  │  │           ┼──┼   │     │
    │ ┌──────────────┐ │  │  │  GATEWAY  │  ┼───┼ ... │
    │ │              │ │  ┼──┼           │  │   │     │
    │ │              ┼─┼──┼  │           │  │   └─────┘
    │ └──────────────┘ │  │  └───────────┘  │
    └──────────────────┘

The actual test starts deploying the client application on the workload
cluster, alongside the IsovalentEgressGatewayPolicy that redirects all
traffic from the client application to the external network via the
gateway nodes. Once the BPF policy maps have been correctly populated,
the client performs a set of requests to the external application, and
asserts that the observed source IP address matches one of the gateways.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>