enterprise/cilium-cli: Detect native routing CIDR in egw-ha IPAM tests
Egress gateway IPAM feature requires a list of egress CIDRs to be used
for IP allocations. These CIDRs must be specified in the IEGP.

The approach used to get a CIDR to allocate addresses from in other
tests is to get the kind cluster native routing CIDR (with `docker
network inspect`) and reserve a small inner CIDR to be used for IPAM.
Finally, the CIDR is passed to the cilium-cli as a command line config
option.  This approach requires to change each workflow where the test
is enabled to run the `docker network inspect` command before running
cilium-cli. This, in turn, means to change a relevant number of workflow
files shared with the OSS counterpart, defeating the mindfulness of
those enterprise tests.

The alternative approach is then to auto-detect the native routing CIDR
before running the test. To do so, we exec into a Cilium pod (that is
running in the host network namespace) to run the following commands:

```
ip --family inet route show default | jq '.[0].dev'
ip --family inet --json addr show <default-interface> | jq '.[0].addr_info.[0].local'
ip --family inet --json addr show <default-interface> | jq '.[0].addr_info.[0].prefixlen'
```

Where <default-interface> is the interface managing the default route,
returned by the first command. The remaining two commands returns the
native routing CIDR, as can be seen from this output from a local kind
cluster:

```
$ kubectl -n kube-system exec -ti ds/cilium -- ip --family inet addr show eth0
23: eth0@if24: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP group default  link-netnsid 0
    inet 172.18.0.5/16 brd 172.18.255.255 scope global eth0
       valid_lft forever preferred_lft forever
```

Where 172.18.0.5/16 is the native routing CIDR.

The egress gateway IPAM tests are then updated to use:

- the user-specified egress CIDRs, if any
- an egress CIDR that is part of the native routing CIDR, if the
  auto-detection is successful
- a default value ("172.18.0.8/30") if the detection fails

Once the native routing CIDR has been detected, we reserve a "/30" CIDR
to be used for IPAM, starting from the eighth address in the native
routing CIDR. This is done to avoid having an IPAM CIDR that may overlap
with the addresses already allocated for the nodes.

Finally, a setup function is registered for the egress gateway IPAM
tests, in order to update the IEGP to be used before applying it.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>