egressgw: introduce standalone egress gateway map
[ v1.15-ce commit bceadb55ed34069b359193ce77ea98b4356df1a5 ]

[ backporter's notes: ignoredELFPrefixes and enterprise_cache.go
  been already removed. Skipped the changed to them ]

Standalone Egress Gateway (SEGW) represents a variant of standard
Egress Gateway designed to be deployed on nodes outside of a K8s
clusters. It is responsible for handling the traffic coming in
from the VXLAN tunnel, redirecting it to the configured egress
interface and masquerading it accordingly.

This commit introduces a new BPF map storing endpoint IP to tunnel
endpoint (and security identity) mappings, in order to handle the
reply path. This approach mimics the backward learning technique
adopted by network switches, and relaxes the need for connecting
to the Cilium datastore (e.g., K8s API Server, KVStore, clustermesh)
to retrieve the ipcache mappings, which are instead directly learned
when the forward direction of the traffic is received from the tunnel.

More specifically, the new BPF map, cilium_egress_gw_standalone_v4,
is of type LRU to automatically handle the removal of old entries,
without requiring an explicit GC process, is keyed by endpoint IP
(i.e., source pod address), and its value consists of the IP of the
tunnel endpoint used to reach the given pod, complemented by the
security identity associated with the source pod. The map is meant to
be populated directly by the Cilium datapath, and possibly accessed
from user-space for debugging purposes only.

The size of the map is configurable via the dedicated
--standalone-egress-gateway-map-max flag (defaults to 2^14 entries,
similarly to the egress gateway policy map).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>