egwha: remove TestEgressGatewayManagerWithoutNodeSelector test.
The TestEgressGatewayManagerWithoutNodeSelector test expects that
setting an empty nodeSelector should yield no active/healthy GW
nodes in the groupSpec.

This is actually not the correct behavior as of v1.14-ce+ and the only
reason that this test is passing is due to a race condition in the test.

This happens when the groupStatus is checked prior to it being updated
by the manager.
Because the reconciliation is async, it is not guarenteed that the
resource
will be updated in time (in fact - it generally isn't, only "flaking"
out about 1 in a few hundred test runs).

If one adds a time.Sleep(time.Second) following setting up test
infra but preceeding assertion this test always fails. ex:

	...

        // Create a new policy without nodeSelector
        k.addPolicy(t, &policyParams{
                name:            "policy-1",
                uid:             policy1UID,
                endpointLabels:  ep1Labels,
                destinationCIDR: destCIDR,
                egressIP:        egressIP1,
        })

	time.Sleep(time.Second) // wait for groupStatus to settle...

        // Operator should select no active / healthy gateways for this policy
        k.assertIegpGatewayStatus(t, gatewayStatus{
                activeGatewayIPs:  []string{},
                healthyGatewayIPs: []string{},
        })

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>