enterprise/cilium-cli: Add NewEnterpriseTestWithoutSetup
Add NewEnterpriseTestWithoutSetup to create an enterprise test without
the "enterprise setup" function already registered.

The "enterprise setup" is in charge of applying policies and deployments
relevant for the enterprise tests. But in case we need to update those
policies and/or deployments just before applying them (e.g: with
information that can be gathered only when Cilium is already running),
we may want to register an additional setup function to be executed
beforehand.

In order to do so, NewEnterpriseTestWithoutSetup is added and Setup is
exported. This way, the caller can compose a test with the correct
sequence of setup functions to update the policies before applying them.

This will be useful in the subsequent commit where a test for egw-ha
IPAM is added. In those tests, we want to detect the native routing CIDR
before setting the IP range to use for allocations.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Fabio Falzoi <fabio.falzoi84@gmail.com>