enterprise/features: Add flag validation test
Test that the flags defined in features.yaml are registered to either
the OSS agent or the operator.

This does not yet validate against the EnterpriseAgent/EnterpriseOperator
as we cannot import "enterprise/daemon" (main package). In a follow-up
we'll refactor the package structure to allow for this. In the meantime
we can set 'skip-flag-check' for enterprise flags.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>