ilb/test: Introduce single node mode
This commit introduces the single node testing mode. The mode is used to
run the ILB e2e tests in environments in which the testing code cannot
just deploy FRR client and LB backend containers on any node (as we do
in the ci-lb).

The mode is intended to be used by an end user to validate their ILB
deployments. In order to run the tests from a host, the following is required:

* The host must have an IP addr which is reachable from LB T1 and T2
  nodes.
* The host must have Docker running.

It is possible to run the tests from a remote node by specifying the
test hosts Docker endpoint. For example:

    DOCKER_HOST="tcp://172.18.0.9:2375" go test ... ./enterprise/test/ilb.

Where 172.18.0.9 is the host IP addr, and :2375 is the dockerd socket.

Once started, the tests deploy containers (FRR client and LB backend
apps) in the host netns, and then do "docker exec $client" to send
requests to the LB nodes which forward them to the apps in the same host
and the same netns. The diagram below illustrates that:

   ┌──────────────────┐
   │   Test host      │
   │                  │    ┌──────────────┐
   │  ┌─────────────┐ │ 2. │              │
   │  │ FRR client  ┼─┼────► T1 ILB node  │
   │  └─────────────┘ │    │              │
   │                  │    └───────┬──────┘
   │  ┌─────────────┐ │            │ 3.
   │  │ HTTP app 1  │ │            │
   │  └─────────────┘ │    ┌───────▼──────┐
   │                  │    │              │
   │  ┌─────────────┐ │ 4. │ T2 ILB node  │
   │  │ HTTP app 2  ◄─┼────┼              │
   │  └─────────────┘ │    └──────────────┘
   │                  │
   │  ┌────────┐      │
   │  │ dockerd│      │
   │  └──▲─────┘      │
   └─────┼────────────┘
      1. │
         │
┌────────┴────────────────────────────────────────────────────────┐
│                                                                 │
│ DOCKER_HOST="tcp://$TEST_HOST:2375" go test --mode=single-node  │
│                                                                 │
└─────────────────────────────────────────────────────────────────┘

Signed-off-by: Martynas Pumputis <m@lambda.lt>

WIP

Signed-off-by: Martynas Pumputis <m@lambda.lt>