lb: decrease default health check interval in E2E tests
Currently, the E2E tests fail quite often when waiting for the BGP or FIB routes.

```
2024-09-27T08:47:34.8333691Z     helpers_testscenario.go:76: timeout reached after 1m0s, last error: prefix 100.64.0.5/32 not found in FIB routes
```

It looks like increasing the wait time from 30s to 60s did help but not completely.

As a next attempt to improve the situation, this commit decrases the
default health check interval of the backendpools from 10s to 5s.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>