lb: export T1 node metrics
this commit introduces 4 new prometheus metrics, related to T1 LB nodes:

- lb_bytes_total
- lb_packets_total
- lb_open_connections_metric
- lb_healthcheck_status

`lb_bytes_total` and `lb_packets_total` are derived from the bpf CT
table, measure the total number of ingress bytes and packets for a
given (service, backend) tuple, and are labelled accordingly with the
service and backend labels they refer to. Some examples:

	cilium_lb_bytes_total{backend="172.18.0.1",service="myservice-1 (100.64.0.0:80/TCP)"} 1000
	cilium_lb_bytes_total{backend="172.18.0.2",service="myservice-1 (100.64.0.0:80/TCP)"} 1400
	cilium_lb_packets_total{backend="172.18.0.1",service="myservice-1 (100.64.0.0:80/TCP)"} 200
	cilium_lb_packets_total{backend="172.18.0.2",service="myservice-1 (100.64.0.0:80/TCP)"} 300

`lb_open_connections_metric` is also derived from the CT table, and
measures the current number of LB open connections:

	cilium_lb_open_connections_metric 42

`lb_healthcheck_status` is derived from the service map and reports the
health status of the backends for a given service, where the value 0
means unhealthy and 1 means healthy:

	cilium_lb_healthcheck_status{backend="172.18.0.1",service="100.64.0.0:80/TCP"} 1
	cilium_lb_healthcheck_status{backend="172.18.0.2",service="100.64.0.0:80/TCP"} 1

Metrics are collected periodically (with the current interval of 5s),
regardless of the scraping time configured for prometheus.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>