lb: remove unsupported hostname types from LBService types
Currently, the documentation states that hostname on a route match
does support prefix based domain matching and also suffix based with
partial wildcards.

But this isn't supported by the implementing regex. And this is the
case for the reason that we we not only use these hostnames when
defining the virtual hosts servernames (that would support all these
cases) but also the filter chain matches for TLS. And this doesn't
support these special cases.

https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/listener/v3/listener_components.proto#envoy-v3-api-msg-config-listener-v3-filterchainmatch

```
server_names
(repeated string) If non-empty, a list of server names (e.g. SNI for TLS protocol) to consider when determining a filter chain match. Those values will be compared against the server names of a new connection, when detected by one of the listener filters.

The server name will be matched against all wildcard domains, i.e. www.example.com will be first matched against www.example.com, then *.example.com, then *.com.

Note that partial wildcards are not supported, and values like *w.example.com are invalid. The value * is also not supported, and server_names should be omitted instead.
```

Therefore, this commit removes the unsupported hostname types from the type documentation.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>