enterprise/features: Feature specifications and gates
Implement support for defining features and feature gates for the features.

The file enterprise/features/features.yaml defines feature maturity levels
and the features. A feature defines a set of flags and helm options that enable
the feature.

The file is embedded into the enterprise cilium-agent and at agent start the
feature gates are checked for each feature that is enabled according to its flags.
If the feature is not approved (and strict feature gate mode is on) then agent will
fail to start. Agent will also check feature gates for operator-specific flags.

Additionally from features.yaml a helm validation template is generated to check the
feature gates at "helm install/upgrade" time to avoid deploying a configuration that
would cause the agent to fail to start.

Signed-off-by: Jussi Maki <jussi@isovalent.com>