enterprise/encryption: Add encryption policy reconciler metrics
This adds two metrics tracking the BPF map reconciler of the encryption
policy subsystem. Both metrics are implemented by the
`reconcilerMetrics` type, which intercepts all operations issued by the
reconciler and updates the metrics accordingly.

The first metric measures the duration it takes the reconciler to
reconcile all pending updates. This is implemented by wrapping the
reconciler operations and tracking the StateDB revision of processed
items. See the inline comment for more details on this approach.

The second metric simply counts the number of operations that errored
out. Note that the reconciler retries any operation that returns an
error, so a non-zero metric does imply a defunct system.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>