lb,test: Add DNS-based backend discovery test
Add a test to make sure the DNS-based backend discovery is working as
expected. It tests basic connectivity as well as the case that multiple
backends are bounded to a single domain name.

It utilizes k8s headless service since Docker doesn't have a good way to
create two containers resolved by the same domain name. That's why we
set Envoy dnsConfig to the ClusterFirstWithHostNet.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>