github: Add cl2 module for encryption policy scale test
This adds a clusterloader2 scale test for the encryption policy
control-plane. It creates a number of deployments (depending on the node
count), each with unique identities:

 - for 1 node -> 10 deployments x 1 namespace
 - for 100 nodes -> 1000 deployments x 1 namespace
 - for 300 nodes -> 1000 deployments x 3 namespaces
 - for 1000 nodes -> 1000 deployments x 10 namespaces

For each node, we in parallel create 25 IsovalentClusterwide-
EncryptionPolicies distributed over the namespaces, with a
subject and peer chosen within a namespace.

The test also gathers a select set of metrics measuring control-plane
latency. This commit in addition also contains a Grafana dashboard
showing various relevant metrics over time.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>