lb: e2e: use HTTP response to assert selected backend
Currently, the log of backend pods is analyzed to assert whether
a request hit the configured backend in the HTTP(S) route E2E tests.

But it would be way easier (and less flaky (due to the access log flushing))
to use the HTTP response of the request. The response already contains the
name of the backend.

Therefore, this commit refactors the tests accordingly.

Suggested-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>