lb: Taint T1 node in the LB CI
In our local development environment, T1 nodes are tainted by
./examples/kubernetes/loadbalancer/lb_configure.sh. This is because the
T1 node doesn't run Cilium CNI, so scheduling (non-hostnetwork)
containers on T1 node leads to the connectivity loss.

This is especially problematic when the CoreDNS is scheduled to the T1
which makes in-cluster DNS cluster impossible.

Currently, the taint is missing on the CI node. Add it.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>