hubble/enterprise: remove outdated //go:build go1.18 tag
This used to be needed to keep Go versioning in sync but has since been
dropped from the Cilium code base since commit 66377a3b2fc1 ("Don't
enforce Go version using //go:build in main.go files"). A similar change
is done for the OSS Hubble CLI in
https://github.com/cilium/cilium/pull/35174.

This build tag also leads to a rather unhelpful error message when
compiling with an outdated Go versions, e.g.

    package github.com/cilium/cilium/hubble: build constraints exclude all Go files in ...

Nothing in the error messages indicates that this is due to an outdated
Go toolchain version and this has been a source of confusion several
times for several developers.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>