enterprise/helm: Require explicit fallbackBehavior for encryption policies
The current implementation of encryption policies selects traffic to be
encrypted via encryption policy. If traffic is not selected by any
policy, then the traffic will be emitted as plaintext.

Future versions of Cilium however might want to invert that logic:
Policies select what traffic to emit as plaintext, and the fallback
behavior is to always encrypt. In particular, since the
"encrypt-by-default" behavior is more secure, it will likely become the
default fallback behavior if implemented.

Therefore, this commit requires that user explicitly specify the
fallback behavior. This allows future versions of Cilium to change the
default fallback behavior without requiring early adopters of the
feature to change their Helm configuration.

The fallback behavior also has to be added to the features.yaml to avoid
issues with the `check-enterprise-features` Makefile target (which would
fail because the fallback behavior is not explicitly set).

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>