lb: de-duplicate Envoy cluster creation
Currently, reconciling an `LBService` creates a dedicated Envoy cluster
for each application route referencing a `LBBackendPool`. The names of
the clusters use a combination of application types and the index of the
route as postfix.

This approach leads to confusing naming, unnecesssary reconciler complexity
(always looping through all application types & their routes) and more
Envoy clusters than necessary.

Therefore, this commit refactores the reconciliation to create an Envoy cluster
for each referenced `LBBackendPool` that is referenced by an `LBService` instance.
The name of the cluster uses the name of the `LBBackendPool` as postfix.

Note: Multiple `LBService` instances referencing the same `LBBackendPool` still result
in different Envoy Clusters!

---

In addition to de-duplicating the number of Envoy Clusters, this commit also has
to introduce an explicit configuration option on the `LBBackendPool`, whether the
health checking should be sent via TLS or not.

Background: This aspect was handled differently for the application type `TLSPassthrough`.
The reconciliation configured the Envoy cluster that is used in the context of a
TLS passthrough route to use an Upstream TLScontext when sending the health check requests
upstream. Now that we no longer have dedicated Envoy clusters per route, it's necessary
that it becomes an explicit configuration, whether health checking should be performed
with TLS or not.

Note: By default Envoy health checking use the same upstream context that is configured
on the Cluster itself. This wasn't working with our test health checking application as
this is either configured to use TLS or not - therefore this behaviour. But in production,
this "implicit" configuration / assumption might be wrong. (The TLS passthrough backend could
also easily serve the health check endpoint without TLS - even though it is expecting the
requests with TLS. This is another good reason why this configuration should be "explicit".

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>