lb: add explicit health check tls config
Currently, health checking (Envoy: T2->Backend) uses the same transport socket
(TLS enabled or disabled) as the one that is used to send application requests to the backend.

The only exception was some special handling for TLS passthrough application where
the health check was implicitly changed to use TLS as the transport socket in TLS
passthrough mode isn't configured to use TLS (as the passed through packets are already
TLS encrypted).

This led to the need to be able to override TLS configuration for backend health checking.
Eventually leading to the decision that it's probably worth to not re-use the same
transport socket as the normal requests to the backends - and instead rely on explicit
configuration for the health checks themself. Aside from being explicit - it's also the
most flexible version where the users can configure health checking TLS to their needs.

With the one drawback that in case of TLS re-encryption to the backend a user has
to configure TLS twice if the health check endpoint usese TLS too.

This commit is introducing explicit TLS configuration for the health check config
on the `LBBackendPool` CRD.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>