lb: Introduce CoreDNS container type
Introduce a new CoreDNS container type. It can be instantiated using
enableCoreDNS helper function of test schenario. It allows to register
A or AAAA record with AddDNSRecords call.

The only use case for now is testing custom DNS resolver and DNS-based
backend discovery. We can extend it to other use cases in the future as
needed.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>