lb: Add test for DNS backend and custom resolver
Add a DNS-based backend discovery test. It leverages the CoreDNS
container introduced in the previous commit. It ensures the DNS
discovery for the multiple backends bounded to the single domain name.

This test works with the single-node mode, but with single backend. This
is because, in DNS-based backends, we cannot specify the port for each
backends individually.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>