lb: fix yq invocation
The yq program in our scripts is executed via a `docker run`. A file to be
patched by yq is bind-mounted inside the container. By default, the user ID
of the process inside a container is 1000 (user yq). If the UID of a person
executing the command is different, then the process not be able to patch
the file. Fix this by executing the yq with the UID/GID of the actual user.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>