lb: validate applications in LBService differently
We only allow one application in LBService and with the number of applications
growing the formula becomes too complex (and the current one actually doesn't
check the full mutual exclusiveness). Correct it and more it more predictable
by using the following formula (say, we have variables x, y, ..., z):

    ( x & !y & ... & !z) ||
    (!x &  y & ... & !z) ||
            ...
    (!x & !y & ... &  z)

Note that the openapi schema allows use of oneOf operator, so theoretically,
it can be used. However, it is not supported by tooling, see a related issue
https://github.com/kubernetes-sigs/controller-tools/issues/461 Once it is
resolved, we can use a simplified validation here.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>