lb: Workaround the issue of no_default_search_domain
Currently the Envoy's no_default_search_domain feature is broken,
because of the issue described in the below issue.

https://github.com/envoyproxy/envoy/issues/33138

Unfortunately, the fix is still not released yet and we need to live
with this issue for a while. As a workaround, we always qualify the
hostname passed to the LBBackendPool so that it works as intended.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>