lb: disable expensive x-request-id header generation by default
Currently, we configure Envoy to generate an x-request-id HTTP
header by default (It's also the default of Envoy' HTTPConnectionManager).

```
generate_request_id
(BoolValue) Whether the connection manager will generate the x-request-id header if it does not exist. This defaults to true. Generating a random UUID4 is expensive so in high throughput scenarios where this feature is not desired it can be disabled.
```

https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/network/http_connection_manager/v3/http_connection_manager.proto

The generation of this UUID4 based x-request-id is expensive and it's recommended
to disable it in high throughput scenarios.

Therefore, this commit is disabling it by default. It's still possible
to opt-in with the config flag.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>