cilium-cli: introduce support for phantom services feature flag
Historically, Phantom Services support was not guarded by a feature
flag, and always enabled. Starting from Cilium v1.16-ce [1,2], instead,
this feature is disabled by default, and can be enabled via the
dedicated enable-phantom-services flag. Hence, let's teach the Cilium
CLI to respect the same flag, and not attempt to run the corresponding
tests when the feature is not enabled.

[1]: 2e499d7be0b0 ("clustermesh: Add option to enable phantom services")
[2]: 2bcd893d5eff ("helm: change default values of limited features")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>