lb: always set service forwarding mode annotation explicitly
Currently, the annotation `service.cilium.io/forwarding-mode` is only set
on the T1 service if the LBService isn't configured to use TCPProxy in T1 mode.
In this case the annotation is set to `dsr`. All other cases are handled by
the implicit forwarding mode default `snat`.

This commit changes this behaviour that the forward mode is always explicitly
set.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>