lb: rename t1 endpoints/svc portname from http to tcp
Currently, the portnames of the T1 endpoint & service are named
`http`. This is not correct for all cases. Therefore, this commit
renames it to `tcp` (lowercase - DNS_LABEL expected).

Note: Even though the field `name` is optional for one port, we still
want to name it - especially if we're going to add UDP in the future.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>