docker: Fix NOSTRIP option in cilium/Dockerfile
In PR https://github.com/cilium/cilium/pull/32660, the MODIFIERS build
argument was added as a replacement for individual build arguments, one
of which was NOSTRIP. There was an uncaught usage of NOSTRIP in cilium's
Dockerfile, which has caused the NOSTRIP modifier to become a no-op.
This commit fixes this usage, allowing for NOSTRIP to be used again.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>