lb,test: Reorganize BGP resource structure
Reorganize the BGP-related resource (ClusterConfig, PeerConfig,
Advertisement, and BFDProfile). Currently, the ClusterConfig is global,
PeerConfig is per-FRR-container, Advertisements are created for each
VIP and per-FRR-container, and the BFDProfile is per-FRR-container.

This commit reorganizes the structure to minimize the config duplication
and clarity. ClusterConfig is still global, PeerConfig and BFDConfig are
per-test-scenario, Advertisement is per-VIP.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>