lb: Implement app-level basic auth reconciliation for HTTPProxy
Implement an app-level basic authentication reconciliation logic.
Note that we implement it for HTTPProxy only in this commit.

The model and ingestor are extended to deal with the new `auth` field.
Currently, only auth.basic is defined, but in the future, we can extend
this to contains jwt or oauth2 for example.

This is the first case that we use the content of the Secret (not the
name of the secret) in the reconciliation logic, so some secret handling
logic needed to be modified. The loadTLSSecrets is modified to return
the map of name => Secret instead of slice of Secret. This map is used
later by ingestor to extract the username-password from the Secret.

The translator generates the listener with a new basic authentication
http filter. During the translation, it converts the clear-text
user-password pairs into the htppasswd format that Envoy basic
authentication filter takes.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>