ilb: Do not allow backends with diff ports for TCPProxy
In order to support T1-only TCPProxy LB service with backends with
different ports, the T1 service translator has to create
corev1.Endpoints with a EndpointSubset for each port. For example,

  Subsets: []slim_corev1.EndpointSubset{
     {
  	Addresses: []slim_corev1.EndpointAddress{
  	  {IP: "1.1.1.1"},
  	},
  	Ports: []slim_corev1.EndpointPort{
  	  {
  	    Name: "tcp",
  	    Port: 8080,
  	  },
  	},
     },
     {
  	Addresses: []slim_corev1.EndpointAddress{
  	  {IP: "2.2.2.2"},
  	},
  	Ports: []slim_corev1.EndpointPort{
  	  {
  	    Name: "tcp",
  	    Port: 8081,
  	  },
  	},
     },
  }

The problem is that kube-controller does not like such Endpoint, and as consequence, it
does not create a corresponding EndpointSlice:

  Error mirroring EndpointSlices for <..> is invalid: ports[1].name: Duplicate value: "tcp"

If we change the Name of the second port, then that port is ignored
by the agent on T1 when correlating Endpoint with Service, as the port
name in the corresponding Service must match the Endpoint's port.

For now return the error when such a service is encountered. In the
future we should either fix the correlation or report the error in the
BackendPools status.

Signed-off-by: Martynas Pumputis <m@lambda.lt>