make: add targets to build and install CEE-only cilium-dnsproxy image in kind
Add the necessary Makefile infrastructure to build, load and install the
cilium-dnsproxy image into a local kind cluster. This simplifies local
testing of changes to the Cilium FQDN HA feature.

Use the following commands to bring up a kind cluster for local
development with Cilium and Cilium DNS Proxy installed:

  make kind
  make kind-image
  make kind-image-enterprise-fqdn-proxy
  make kind-fqdnha-install-cilium

Signed-off-by: Tobias Klauser <tobias@isovalent.com>