egressgwha: Avoid useless check for previous egress IPs
The egress IP allocation algorithm used for egress gateway IPAM strives
to keep the same egress IP allocated previously when an egress gateway
remains active between two consecutive changes in the IEGP. To do that,
we first check the previous allocations map and, if the current gateway
IP is a valid key in that map, we first try to reserve the previous
allocated egress IP instead or getting the next one available.  Once all
the previously active gateways have been checked, we can proceed with
the usual allocation logic.

This commit rmeoves a useless further check against the previous
allocations map when we already know that the current active gateway IP
does not have any previous allocation.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>