feat: Add a lifecycle operator for Cilium
A replacement of cilium-ee-olm that leads to an improved deployment experience
on OpenShift and later on other platforms combined with a lighter, less time
consuming release process.
It aims at setting the foundations for advanced capabilities like:
changes and upgrades readiness for a safer user experience,
controlling out-of-band changes, environment aware auto-configuration.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>